requirements.txt
fastapi
uvicorn
google-cloud-texttospeech
pydantic